pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]

[project]
description = ""
dynamic = ["version"]
name = "takeout-maps"

authors = [{name = "Mahdi Lamb", email = "mahdilamb@gmail.com"}]
dependencies = [
  "loguru",
  "fastapi",
  "uvicorn[standard]",
  "pydantic-stream@git+https://github.com/mahdilamb/pydantic-stream",
  "ijson",
  "fitbit-web@git+https://github.com/mahdilamb/fitbit-web",
  "sqlalchemy",
  "tqdm",
]

[project.optional-dependencies]
all = ["takeout-maps[test,dev]"]
dev = [
  "xsd-to-pydantic@git+https://github.com/mahdilamb/xsd-to-pydantic",
  "pydantic-to-typescript@git+https://github.com/mahdilamb/pydantic-to-typescript@changes-for-new-pydantic",
]
test = [
  "black",
  "isort",
  "mypy",
  "bandit",
  "pytest",
  "pytest-cov",
  "docformatter",
  "pydocstyle",
  "types-requests",
]
[tool.setuptools]
packages = ["takeout_maps"]
[tool.setuptools.dynamic]
version = {attr = "takeout_maps.__version__"}

[tool.bandit]
exclude_dirs = ["tests"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.docformatter]
black = true
recursive = true

[tool.mypy]
plugins = "pydantic.mypy"

[tool.pytest.ini_options]
minversion = "6.0"
python_files = [
  "test_*.py",
]
python_functions = ["*_test"]
testpaths = [
  "tests",
]
